Block ASCII NUL and `\` in usernames.
We already blocked `/`.

We did not block NUL, but there was no vulnerability: we wouldn't have
been looking at what we thought we were looking at, and passing bad data
to user lookup in passwd seems bad, and it's generally dodgy anyway, so
we now explicitly disallow it on principle.

On Windows, not blocking `\` would be a security problem, but we disavow
our suitability for running on Windows.  Block it anyway.

For good measure, make sure we additionally block whatever the current
OS's path separator is, if it's not one of those.

Document some of this in AttackSurface.md and explicitly note that we
_do_ allow `@` and non-roman-alphabet UTF-8.