Docs: build static; FreeBSD update steps
The build static command-line should be in the generic install steps
for convenience, not just in the example for the OS we describe.

Set a better example in documentation habits, by also putting a slightly
massaged form of my upgrade-step checklist into the public
documentation, to help form the core of the docs should anyone else ever
actually use this.