Switch to SetDeadline(past) to shut down listener
Something Brady pointed out a few weeks back and which made me rather
happier than a non-programmer might understand: in Go, calling
SetDeadline() on a bound socket used for listening, to set a time in the
past, causes go-routines in other threads to immediately exit.

Per the docs:
> The deadline applies to all future and pending I/O, not just the immediately following call to Read or Write.

This is go-routine safe.

This means we don't need to repeatedly wake up, just to see if we should
exit, and can instead sleep and even get paged out.

Keep the `-listen.at-a-time` option but make it a no-op.

Bump version number: 0.1.4