Various small cleanups N years later
**One logging message change worthy of note if you have log alerts set up.**

Read through all the code, mostly fixed up comments to better explain to
later-me what earlier-me was thinking.

Added `-log.level` flag to let the logging levels be controlled; likely
just an inadvertent omission.  Give it `help` or empty to list the
available values (as a log line).  A rare case where we exit immediately
without tight-loop protection.

Switch home-dir stat (without passwd) to be Lstat, as comments seem to
imply I thought I was doing.  Log better the filetype if passwd lists a
non-dir as home-dir.

The log-message when there's a file ownership mismatch has changed; we
now check more sensibly at filename stat time, and log with fewer caps
there, but then when at file-descriptor stat if there's a conflict,
we're more sure it's an attack, so be more accusatory in our ALL CAPS
message.

Count the "change deadline to terminate accept" as something actively
using a listener; it turns out we're not using this WaitGroup and should
consider removing it, so there's probably nothing nuking the structure
out during shutdown while still in use, but this is slightly more
correct.