Double-check FS notify on alias file
With a K8S configmap, we get a file-gone notification when the file is
replaced, but the file does still exist.  The kernel will have removed
its watch automatically (and the fsnotify lib cleaned up its state), but
we never get a corresponding Create notification for the replacement
file.

Before, we just manually removed the watch, which was a harmless no-op.
Now, we stat() to see if the file exists and only if it definitively
does not, do we let it ride.  Otherwise, we re-add the watch.

I'm not comfortable with the race here, between file ArevN being deleted
and ArevN′ replacing it, in those scenarios, but I'm just going to have
to hope that for a visible window like that, we will get the creation
notification.