Switch error handling to Go 1.13+ approach
Go 1.15+ has changes to stdlib error returns, starting with tls.Conn() wrapping
the error type.  To be resilient against future such changes, use `errors.As()`
to check an error chain to see if it holds an error of the given type, rather
than attempting to just interface cast the top-level returned error.

The `errors.As()` function was introduced with Go 1.13, and with this week's
release of Go 1.15, the minimum supported upstream is now Go 1.14.