Become staticcheck clean; fix signal notification
staticcheck complained about 4 extraneous returns, which IMO is stylistic but
it's a common enough tool that I might as well silence it.

It also caught a real bug: signal.Notify *MUST* get a _buffered_ channel, since
the signal package won't block when sending to it.

At this point I'm not adding staticcheck to Circle CI.  Punting for a later
switch to something Actions-style.